% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_dnorm_area}
\alias{geom_dnorm_area}
\title{Normal Distribution Density Function}
\usage{
geom_dnorm_area(xmin, xmax, mean = 0, sd = 1, n = 101, ...)
}
\arguments{
\item{xmin}{Minimum value for the x-axis}

\item{xmax}{Maximum value for the x-axis}

\item{mean}{Mean of the normal distribution}

\item{sd}{Standard deviation of the normal distribution}

\item{n}{Number of points to interpolate along}

\item{...}{Arguments passed to \code{\link{geom_ribbon}}.}
}
\description{
Plot area under the normal distribution density curve.
}
\examples{
geom_dnorm_area(-2, 2)
mu <- 1
sigma <- 2
geom_dnorm_area(0, 2, mean = mu, sd = sigma)
}
\seealso{
Other Plots.of.the.normal.distribution: \code{\link{geom_dnorm_line}};
  \code{\link{normal_area_plot_p}},
  \code{\link{normal_area_plot_q}};
  \code{\link{normal_plot}};
  \code{\link{normal_tail_plot_p}},
  \code{\link{normal_tail_plot_q}}
}

