% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_dnorm_line}
\alias{geom_dnorm_line}
\title{Area Under the Normal Distribution Density Function}
\usage{
geom_dnorm_line(xmin, xmax, mean = 0, sd = 1, n = 101, ...)
}
\arguments{
\item{xmin}{Minimum value for the x-axis}

\item{xmax}{Maximum value for the x-axis}

\item{mean}{Mean of the normal distribution}

\item{sd}{Standard deviation of the normal distribution}

\item{n}{Number of points to interpolate along}

\item{...}{Arguments passed to \code{\link{geom_line}}.}
}
\description{
Plot the normal distribution density function.
}
\examples{
geom_dnorm_line(-4, 4)
mu <- 2
sigma <- 3
geom_dnorm_line(mu - 3 * sigma, mu + 3 * sigma, mean = mu, sd = sigma)
}
\seealso{
Other Plots.of.the.normal.distribution: \code{\link{geom_dnorm_area}};
  \code{\link{normal_area_plot_p}},
  \code{\link{normal_area_plot_q}};
  \code{\link{normal_plot}};
  \code{\link{normal_tail_plot_p}},
  \code{\link{normal_tail_plot_q}}
}

