% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_dt_line}
\alias{geom_dt_line}
\title{Student's t Distribution Density Function}
\usage{
geom_dt_line(df, xmin, xmax, mean = 0, scale = 1, n = 101, ...)
}
\arguments{
\item{df}{Degrees of freedom of the Student's t distribution.}

\item{xmin}{Minimum value for the x-axis}

\item{xmax}{Maximum value for the x-axis}

\item{mean}{Location of the Student's t distribution}

\item{scale}{Scale of the of the Student's t distribution}

\item{n}{Number of points to interpolate along}

\item{...}{Arguments passed to \code{\link{geom_line}}.}
}
\description{
Plot the Student's t distribution density function.
}
\seealso{
Other Plots.of.the.t.distribution: \code{\link{geom_dt_area}};
  \code{\link{students_t_area_plot_p}},
  \code{\link{students_t_area_plot_q}};
  \code{\link{students_t_plot}};
  \code{\link{students_t_tail_plot_p}},
  \code{\link{students_t_tail_plot_q}}
}

