% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{normal_area_plot_q}
\alias{normal_area_plot_p}
\alias{normal_area_plot_q}
\title{Normal Distribution Density Plot with Area Shaded}
\usage{
normal_area_plot_q(qmin, qmax, mean = 0, sd = 1, max.sd = 4,
  area_opts = list(), ...)

normal_area_plot_p(p, mean = 0, sd = 1, ...)
}
\arguments{
\item{qmin}{Lower quantile of the shaded area}

\item{qmax}{Upper quantile of the shaded area}

\item{mean}{Location}

\item{sd}{Scale}

\item{max.sd}{The x-axis of the plot is \code{mean + c(-1, 1) * max.sd * sd }.}

\item{area_opts}{Additional arguments passed to \code{\link{geom_dnorm_area}}.}

\item{...}{Arguments passed to \code{\link{normal_plot}}.}

\item{p}{Probabilities in the shaded area. The area is centered at the median
of the distribution.}
}
\description{
Plot of the Normal distribution density function, with typical defaults used
in introducing normal distributions, and an internal area shaded.
}
\details{
\code{normal_area_plot_q} plots the density function and shades the area between \code{qmin}
and \code{qmax}. \code{normal_area_plot_p} plots the density function and shades
the area centered at the median such that \eqn{P(lb \leq X \leq ub) = p}.
}
\seealso{
Other Plots.of.the.normal.distribution: \code{\link{geom_dnorm_area}};
  \code{\link{geom_dnorm_line}}; \code{\link{normal_plot}};
  \code{\link{normal_tail_plot_p}},
  \code{\link{normal_tail_plot_q}}
}

