% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{normal_plot}
\alias{normal_plot}
\title{Normal Distribution Density Plot}
\usage{
normal_plot(mean = 0, sd = 1, max.sd = 4, n = 101, digits = 2, ...)
}
\arguments{
\item{mean}{Mean of the normal distribution}

\item{sd}{Standard deviation of the normal distribution}

\item{max.sd}{The x-axis of the plot is \code{mean + c(-1, 1) * max.sd * sd }.}

\item{n}{Number of points to interpolate along}

\item{digits}{Number of digits to display in x-axis breaks.}

\item{...}{Arguments passed to \code{\link{geom_dnorm_line}}.}
}
\description{
Plot of the normal distribution density function, with typical defaults used
in introducing normal distributions.
}
\seealso{
Other Plots.of.the.normal.distribution: \code{\link{geom_dnorm_area}};
  \code{\link{geom_dnorm_line}};
  \code{\link{normal_area_plot_p}},
  \code{\link{normal_area_plot_q}};
  \code{\link{normal_tail_plot_p}},
  \code{\link{normal_tail_plot_q}}
}

