% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{normal_tail_plot_q}
\alias{normal_tail_plot_p}
\alias{normal_tail_plot_q}
\title{Normal Distribution Density Plot with Tail Areas}
\usage{
normal_tail_plot_q(q, mean = 0, sd = 1, lower.tail = TRUE,
  two.sided = FALSE, area_opts = list(), max.sd = 4, ...)

normal_tail_plot_p(p, mean = 0, sd = 1, lower.tail = TRUE,
  two.sided = FALSE, ...)
}
\arguments{
\item{q}{Quantile to use for the tails}

\item{mean}{Mean of the normal distribution}

\item{sd}{Standard deviation of the normal distribution}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P(X \leq x)},
else probabilities are \eqn{P(X > x)}.}

\item{two.sided}{logical; if \code{TRUE}, two-sided tails. The tails are are
at \code{c(-q, q)} if qunatiles are used, and at the quantiles such that
\eqn{P(X \leq x) = p} and \eqn{P(X > x)} if probabilities are used.}

\item{area_opts}{list. Arguments passed to \code{\link{geom_dnorm_area}}.}

\item{max.sd}{The x-axis of the plot is \code{mean + c(-1, 1) * max.sd * sd }.}

\item{...}{Arguments passed to \code{\link{geom_dnorm_line}}.}

\item{p}{Probabilities of the tail areas}
}
\description{
Plot of the normal distribution density function, with typical defaults used
in introducing normal distributions, and tail areas shaded.
}
\seealso{
Other Plots.of.the.normal.distribution: \code{\link{geom_dnorm_area}};
  \code{\link{geom_dnorm_line}};
  \code{\link{normal_area_plot_p}},
  \code{\link{normal_area_plot_q}};
  \code{\link{normal_plot}}
}

