% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dist_density}
\alias{plot_dist_density}
\alias{plot_dist_density_p}
\title{Plot probability ditribution}
\usage{
plot_dist_density(dist, xmin = NULL, xmax = NULL, area = NULL,
  tails = FALSE, line_opts = list(), area_opts = list(), n = 101, ...)

plot_dist_density_p(dist, xmin = NULL, xmax = NULL, p = NULL,
  tails = FALSE, line_opts = list(), area_opts = list(), n = 101, ...)
}
\arguments{
\item{dist}{Distribution name. Used for functions \code{pdist}, \code{ddist},
\code{qdist}.}

\item{xmin}{Minimum value on the x-axis scale}

\item{xmax}{Maximum value on the x-axis scale}

\item{area}{numeric vector of length 2. Minimum and maximum values of an area
under the curve to shade.}

\item{tails}{logical. Shade areas in the tails}

\item{line_opts}{list. Arguments passed to \code{\link{geom_line}}.}

\item{area_opts}{list. Arguments passed to \code{\link{geom_ribbon}}.}

\item{n}{Number of points to interpolate over}

\item{...}{Arguments passed to the distribution functions.}

\item{p}{Probability area to shade.}
}
\description{
Plot probability ditribution
}
\examples{
plot_dist_density("norm", -4, 4, area = c(-2, 2))
plot_dist_density("norm", -4, 4, area = c(-2, 2), tails = TRUE)
plot_dist_density_p("norm", -4, 4, p = 0.95)
plot_dist_density_p("norm", -4, 4, p = 0.95, tails = TRUE)
}

