\docType{package}
\name{pols208}
\alias{pols208}
\alias{pols208-package}
\title{POLS208}
\description{
<<<<<<< HEAD
Code and data for POLS 208: Political Methods, Emory
University, Spring 2014.
=======
  Code and data for POLS 208: Political Methods, Emory
  University, Spring 2014.
>>>>>>> b2abc13b6855f2a527d71e4c7f4fb611ee27751a
}

