% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{students_t_area_plot_q}
\alias{students_t_area_plot_p}
\alias{students_t_area_plot_q}
\title{Student's t Distribution Density Plot with Area Shaded}
\usage{
students_t_area_plot_q(qmin, qmax, df, mean = 0, scale = 1, max.scale = 4,
  area_opts = list(), ...)

students_t_area_plot_p(p, df, mean = 0, scale = 1, ...)
}
\arguments{
\item{qmin}{Lower quantile of the shaded area}

\item{qmax}{Upper quantile of the shaded area}

\item{df}{Degrees of freedom}

\item{mean}{Location}

\item{scale}{Scale}

\item{max.scale}{The x-axis of the plot is \code{mean + c(-1, 1) * max.sd * sd }.}

\item{area_opts}{Additional arguments passed to \code{\link{geom_dnorm_area}}.}

\item{...}{Arguments passed to \code{\link{students_t_plot}}.}

\item{p}{Probabilities in the shaded area. The area is centered at the median
of the distribution.}
}
\description{
Plot of the Student's t distribution density function, with typical defaults used
in introducing Student's t distributions, and an internal area shaded.
}
\details{
\code{students_t_area_plot_q} plots the density function and shades the area between \code{qmin}
and \code{qmax}. \code{students_t_area_plot_p} plots the density function and shades
the area centered at the median such that \eqn{P(lb \leq X \leq ub) = p}.
}
\seealso{
Other Plots.of.the.t.distribution: \code{\link{geom_dt_area}};
  \code{\link{geom_dt_line}};
  \code{\link{students_t_plot}};
  \code{\link{students_t_tail_plot_p}},
  \code{\link{students_t_tail_plot_q}}
}

