% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{students_t_plot}
\alias{students_t_plot}
\title{Student's t Density Plot}
\usage{
students_t_plot(df, mean = 0, scale = 1, max.scale = 4, n = 101,
  digits = 2, ...)
}
\arguments{
\item{df}{Degrees of freedom}

\item{mean}{Location}

\item{scale}{Scale}

\item{max.scale}{The x-axis of the plot is
\code{mean + c(-1, 1) * max.scale * scale }.}

\item{n}{Number of points to interpolate along}

\item{digits}{Number of digits to display in x-axis breaks.}

\item{...}{Arguments passed to \code{\link{geom_dt_line}}.}
}
\description{
Plot of the Student's t distribution density function, with typical defaults used
in introducing t distributions.
}
\seealso{
Other Plots.of.the.t.distribution: \code{\link{geom_dt_area}};
  \code{\link{geom_dt_line}};
  \code{\link{students_t_area_plot_p}},
  \code{\link{students_t_area_plot_q}};
  \code{\link{students_t_tail_plot_p}},
  \code{\link{students_t_tail_plot_q}}
}

