% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{students_t_tail_plot_q}
\alias{students_t_tail_plot_p}
\alias{students_t_tail_plot_q}
\title{Student's t Density Plot with Tail Areas}
\usage{
students_t_tail_plot_q(q, df, mean = 0, scale = 1, lower.tail = TRUE,
  two.sided = FALSE, area_opts = list(), max.scale = 4, ...)

students_t_tail_plot_p(p, df, mean = 0, scale = 1, lower.tail = TRUE,
  two.sided = FALSE, ...)
}
\arguments{
\item{q}{Quantile to use for the tails}

\item{df}{Degrees of freedom}

\item{mean}{Location}

\item{scale}{Scale}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P(X \leq x)},
else probabilities are \eqn{P(X > x)}.}

\item{two.sided}{logical; if \code{TRUE}, two-sided tails. The tails are are
at \code{c(-q, q)} if quantiles are used, and at the quantiles such that
\eqn{P(X \leq x) = p} and \eqn{P(X > x)} if probabilities are used.}

\item{area_opts}{list. Arguments passed to \code{\link{geom_dt_area}}}

\item{max.scale}{The x-axis of the plot is
\code{mean + c(-1, 1) * max.scale * scale }.}

\item{...}{Arguments passed to \code{\link{geom_dt_line}}.}

\item{p}{Probabilities of the tail areas}
}
\description{
Plot of the Student's t distribution density function, with typical defaults used
in introducing normal distributions, and tail areas shaded.
}
\seealso{
Other Plots.of.the.t.distribution: \code{\link{geom_dt_area}};
  \code{\link{geom_dt_line}};
  \code{\link{students_t_area_plot_p}},
  \code{\link{students_t_area_plot_q}};
  \code{\link{students_t_plot}}
}

